requirements.txt
# Core dependencies
python-dotenv==1.0.0
anthropic==0.39.0

# Browser automation
playwright==1.48.0

# Data handling
pydantic==2.9.2
pyyaml==6.0.2

# Utilities
rich==13.9.2  # For better console output
